refactor(codesandbox): split monolithic lib.rs into focused modules
Break the 2,553-line lib.rs into logical modules:
- types.rs: constants, API response types, SandboxState, utilities
- client.rs: CodeSandboxClient HTTP client (Management + Pint APIs)
- state.rs: session secrets-backed state persistence helpers
- tools/sandbox.rs: create, list, manage lifecycle tools
- tools/exec.rs: command execution and status tools
- tools/files.rs: file read/write and workspace download tools
- tools/git.rs: git clone with GitHub auth support

All 72 unit tests + 6 integration tests pass. Added new tests for
edge cases (empty paths, deserialization defaults, pint empty body).
Added README with module overview and architecture links.
